Move LoggingCog to log.py
